commit builtin exit, fix ft_atoi

+++ libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: okumurahyu <okumurahyu@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/10 12:11:37 by okumurahyu        #+#    #+#             */
-/*   Updated: 2022/02/06 22:31:27 by okumurahyu       ###   ########.fr       */
+/*   Updated: 2022/03/18 17:51:17 by okumurahyu       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int			i;
 	int			sign;
